user access token and login functionality implement

+++ backend/src/utils/cloudinary.ts
@@ -1,15 +1,20 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+
+dotenv.config({
+    path: "./.env",
+});
 
 const cloudinaryConfig = {
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME!,
-    api_key: process.env.CLOUDINARY_API_KEY!,
-    api_secret: process.env.CLOUDINARY_API_SECRET!,
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 };
 
 cloudinary.config(cloudinaryConfig);
 
-const uploadOnCloudinary = async (filePath: string) => {
+const uploadOnCloudinary = async (filePath: string | undefined) => {
     try {
         if(!filePath) {
             throw new Error("File path is required");
@@ -22,10 +27,9 @@ const uploadOnCloudinary = async (filePath: string) => {
         fs.unlinkSync(filePath); // Clean up the file after upload
         return result;
     } catch (error: any) {
-        fs.unlinkSync(filePath); // Clean up the file after upload
+        if (filePath) fs.unlinkSync(filePath); // Clean up the file after upload
         console.error("Error uploading file to Cloudinary:", error);
         throw error;
-        return null;
     }
 }
